requirements.txt
pygame-ce>=2.3.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-timeout>=2.0.0
pyinstaller>=6.0.0
cython>=3.0.0
setuptools>=68.0.0
numpy>=1.24.0

